+++ BuyItPlatform.GatewayApi/Controllers/UserRatingController.cs
@@ -23,7 +23,7 @@ namespace BuyItPlatform.GatewayApi.Controllers
         public async Task<IActionResult> GetUserRating(string targetUserId)
         {
             var apiResult = await userRatingService.GetUserRating<UserRatingResponseDto>(targetUserId);
-            return StatusCode((int)apiResult.StatusCode, apiResult);
+            return StatusCode(apiResult.StatusCode, apiResult);
         }
 
         [HttpGet]
@@ -31,7 +31,7 @@ namespace BuyItPlatform.GatewayApi.Controllers
         public async Task<IActionResult> DeleteOfferedRatings(string targetUserId)
         {
             var apiResult = await userRatingService.DeleteOfferedRatings<object>(targetUserId);
-            return StatusCode((int)apiResult.StatusCode, apiResult);
+            return StatusCode(apiResult.StatusCode, apiResult);
         }
 
         [HttpPost]
@@ -39,7 +39,7 @@ namespace BuyItPlatform.GatewayApi.Controllers
         public async Task<IActionResult> RateUser([FromBody] UserRatingRequestDto ratingRequest)
         {
             var apiResult = await userRatingService.RateUser<object>(ratingRequest);
-            return StatusCode((int)apiResult.StatusCode, apiResult);
+            return StatusCode(apiResult.StatusCode, apiResult);
         }
     }
 }
